Updates for new features, needs fixing. fixed contract id and quantity to be ints universally.

+++ src/broker_api/models.py
@@ -46,9 +46,9 @@ class Order(BaseModel):
 
 class Fill(BaseModel):
     order_id: int
-    contract_id: float
+    contract_id: int
     time: datetime
-    quantity: float
+    quantity: int
     price: float
     side: OrderSide
